OccasionalCriminalPoints/OccasionalCriminalPoints/pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord-py-interactions==5.13.2",
    "discord-py>=2.5.2",
    "python-dotenv>=1.1.0",
]
